#13 Minor refactoring for making projects-config more standalone + Minor other smaller refactorings
